roads and feature clear

+++ Civilization/src/main/java/Model/Units/Combat/Combat.java
@@ -41,7 +41,7 @@ public class Combat extends Unit {
         this.fortifiedTurnCount = fortifiedTurnCount;
     }
 
-    public int getMovementsLeft() {
+    public double getMovementsLeft() {
         return movementsLeft;
     }
